+++ site/src/hooks/recovery/useRecoverableRecoveriesData.js
@@ -1,7 +1,7 @@
 import { useQuery } from "@apollo/client";
 import { GET_RECOVERABLE_RECOVERIES } from "../../services/gql/recovery";
 import { LIST_DEFAULT_PAGE_SIZE } from "../../utils/constants";
-import { useRecoverableParams } from "../../utils/hooks/recovery/useRecoverableParams";
+import { useRecoverableParams } from "./useRecoverableParams";
 
 export function useRecoverableRecoveriesData() {
   const { address, height, page = 1 } = useRecoverableParams();
